add user input

+++ Main.java
@@ -1,3 +1,4 @@
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
@@ -9,15 +10,34 @@ public class Main{
 	Grid g = new Grid();
 	frame.setSize(WIDTH, HEIGHT);
 	frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+	g.addKeyListener(new KeyListener(){
+		public void keyPressed(KeyEvent e){}
+		public void keyReleased(KeyEvent e){}
+		public void keyTyped(KeyEvent e){
+		    if(e.getKeyChar() == 'e'){
+			g.rotate(1);
+		    }else if(e.getKeyChar() == 'q'){
+			g.rotate(3);
+		    }else if(e.getKeyChar() == 'a'){
+			g.shift(-1);
+		    }else if(e.getKeyChar() == 'd'){
+			g.shift(1);
+		    }
+		}
+	    });
 	frame.add(g);
 	frame.pack();
 	frame.setVisible(true);
-	new Timer(1000, new ActionListener(){
+	new Timer(50, new ActionListener(){
+		private int state = 0;
 		public void actionPerformed(ActionEvent e){
-		    g.drop();
-		    g.rotate(3);
-		    g.repaint();
+		    if(++state == 10){
+			g.drop();
+			g.repaint();
+			state = 0;
+		    }
 		}
 	    }).start();
+
     }
 }
